Chore: Migrate getSlashCommands to TS (#3711)
* migrate getSlashCommands to TS

* use sdk and update getSlashCommands

* minor tweak

* Remove implicit anys
