+++ app/definitions/ISlashCommand.ts
@@ -9,4 +9,8 @@ export interface ISlashCommand {
 	appId?: string;
 }
 
+export interface ISlashCommandResult extends ISlashCommand {
+	command: string;
+}
+
 export type TSlashCommandModel = ISlashCommand & Model;
